move the nmea
